ctuscenes changeworld fix door animation and sound added tripollo physicsweapona attached to a bone

+++ trunk/OUAN/OUAN/Src/Game/GameObject/GameObjectBridge.cpp
@@ -99,7 +99,7 @@ void GameObjectBridge::changeToWorld(int newWorld, double perc)
 	}
 
 	std::string currentAnimName=mRenderComponentEntity->getCurrentAnimationName();
-	float currentAnimLen=mRenderComponentEntity->getCurrentAnimationLength();
+	double currentAnimLen=mRenderComponentEntity->getCurrentAnimationLength();
 	if(!mRenderComponentEntity->getCurrentAnimation()) return;
 
 	switch(newWorld)
